abstract a common interface for planner
