+++ fe/fe-core/src/test/java/org/apache/doris/analysis/SelectStmtTest.java
@@ -20,7 +20,7 @@ package org.apache.doris.analysis;
 import org.apache.doris.common.AnalysisException;
 import org.apache.doris.common.Config;
 import org.apache.doris.common.util.Util;
-import org.apache.doris.planner.Planner;
+import org.apache.doris.planner.OriginalPlanner;
 import org.apache.doris.qe.ConnectContext;
 import org.apache.doris.qe.VariableMgr;
 import org.apache.doris.utframe.DorisAssert;
@@ -562,28 +562,28 @@ public class SelectStmtTest {
     @Test
     public void testSelectHintSetVar() throws Exception {
         String sql = "SELECT sleep(3);";
-        Planner planner = dorisAssert.query(sql).internalExecuteOneAndGetPlan();
+        OriginalPlanner planner = (OriginalPlanner) dorisAssert.query(sql).internalExecuteOneAndGetPlan();
         Assert.assertEquals(VariableMgr.getDefaultSessionVariable().getQueryTimeoutS(),
                 planner.getPlannerContext().getQueryOptions().query_timeout);
 
         sql = "SELECT /*+ SET_VAR(query_timeout = 1) */ sleep(3);";
-        planner = dorisAssert.query(sql).internalExecuteOneAndGetPlan();
+        planner = (OriginalPlanner) dorisAssert.query(sql).internalExecuteOneAndGetPlan();
         Assert.assertEquals(1, planner.getPlannerContext().getQueryOptions().query_timeout);
 
         sql = "select * from db1.partition_table where datekey=20200726";
-        planner = dorisAssert.query(sql).internalExecuteOneAndGetPlan();
+        planner = (OriginalPlanner) dorisAssert.query(sql).internalExecuteOneAndGetPlan();
         Assert.assertEquals(VariableMgr.getDefaultSessionVariable().getMaxExecMemByte(),
                 planner.getPlannerContext().getQueryOptions().mem_limit);
 
         sql = "select /*+ SET_VAR(exec_mem_limit = 8589934592) */ poi_id, count(*) from db1.partition_table "
                 + "where datekey=20200726 group by 1";
-        planner = dorisAssert.query(sql).internalExecuteOneAndGetPlan();
+        planner = (OriginalPlanner) dorisAssert.query(sql).internalExecuteOneAndGetPlan();
         Assert.assertEquals(8589934592L, planner.getPlannerContext().getQueryOptions().mem_limit);
 
         int queryTimeOut = dorisAssert.getSessionVariable().getQueryTimeoutS();
         long execMemLimit = dorisAssert.getSessionVariable().getMaxExecMemByte();
         sql = "select /*+ SET_VAR(exec_mem_limit = 8589934592, query_timeout = 1) */ 1 + 2;";
-        planner = dorisAssert.query(sql).internalExecuteOneAndGetPlan();
+        planner = (OriginalPlanner) dorisAssert.query(sql).internalExecuteOneAndGetPlan();
         // session variable have been changed
         Assert.assertEquals(1, planner.getPlannerContext().getQueryOptions().query_timeout);
         Assert.assertEquals(8589934592L, planner.getPlannerContext().getQueryOptions().mem_limit);
